Incorporate fuel type and station selected in request (#1725)
Added fuel type id to request (not being used right now though - just there).
Changed structure of request to combine fuel type id and "station selected" under planning area.
Split up tests to make running different scenarios from the command line a bit simpler.
