+++ api/app/schemas/hfi_calc.py
@@ -135,6 +135,14 @@ class HFIWeatherStationsResponse(BaseModel):
     fire_centres: List[FireCentre]
 
 
+class StationInfo(BaseModel):
+    """ Information about a station, including its code, name, and elevation. """
+    station_code: int
+    selected: bool
+    # fuel_type_id matches to table fuel_types.id
+    fuel_type_id: int
+
+
 class HFIResultRequest(BaseModel):
     """
     Request that contains inputs necessary for calculating HFI.
@@ -147,7 +155,8 @@ class HFIResultRequest(BaseModel):
     selected_prep_date: Optional[date]
     start_date: Optional[date]
     end_date: Optional[date]
-    selected_station_code_ids: List[int]
+    # Each planning area has a list of stations
+    planning_area_station_info: Mapping[int, List[StationInfo]]
     selected_fire_center_id: int
     # Mapping from planning area id to a map of FireStartRanges.
     planning_area_fire_starts: Mapping[int, List[FireStartRange]]
@@ -162,7 +171,7 @@ class HFIResultResponse(BaseModel):
     selected_prep_date: date
     start_date: date
     end_date: date
-    selected_station_code_ids: List[int]
+    planning_area_station_info: Mapping[int, List[StationInfo]]
     selected_fire_center_id: int
     planning_area_hfi_results: List[PlanningAreaResult]
     # Mapping from planning area id to a map of FireStartRanges
