+++ api/app/tests/hfi/conftest.py
@@ -1,0 +1,17 @@
+import json
+from pytest_bdd import then
+
+
+@then("the response is <response_json>")
+def then_response(result, response_json: dict):
+    """ Check entire response """
+    if response_json is not None:
+        print('actual:\n{}'.format(json.dumps(result['response'].json(), indent=4)))
+        print('expected:\n{}'.format(json.dumps(response_json, indent=4)))
+        assert result['response'].json() == response_json, result['filename']
+
+
+@then("the response status code is <status_code>")
+def then_status(result, status_code: int):
+    """ Check response status code """
+    assert result['response'].status_code == status_code, result['filename']
