+++ api/app/tests/hfi/test_hfi_endpoint_request_load_no_date_specified.json
@@ -1,7 +1,13 @@
 {
-    "selected_station_code_ids": [
-        230
-    ],
+    "planning_area_station_info": {
+        "1": [
+            {
+                "station_code": 230,
+                "selected": true,
+                "fuel_type_id": 1
+            }
+        ]
+    },
     "selected_fire_center_id": 1,
     "planning_area_fire_starts": {}
 }
