+++ api/app/tests/hfi/test_hfi_endpoint_request_save_invalid.json
@@ -1,8 +1,14 @@
 {
     "selected_prep_date": "2022-02-11",
-    "selected_station_code_ids": [
-        230
-    ],
+    "planning_area_station_info": {
+        "1": [
+            {
+                "station_code": 230,
+                "selected": true,
+                "fuel_type_id": 1
+            }
+        ]
+    },
     "selected_fire_center_id": 1,
     "planning_area_fire_starts": {},
     "save": true
