+++ api/app/tests/hfi/test_hfi_endpoint_request_save_valid.json
@@ -2,9 +2,15 @@
     "selected_prep_date": "2020-05-21",
     "start_date": "2020-05-21",
     "end_date": "2020-05-26",
-    "selected_station_code_ids": [
-        230
-    ],
+    "planning_area_station_info": {
+        "1": [
+            {
+                "station_code": 230,
+                "selected": true,
+                "fuel_type_id": 1
+            }
+        ]
+    },
     "selected_fire_center_id": 1,
     "planning_area_fire_starts": {},
     "save": true
